started delivery class

+++ src/CyclistType.java
@@ -2,13 +2,13 @@
 public enum CyclistType {
 	BEGINNER(10), SPORTIF(15), EXPERT(20); 
 	
-	private int speed;
+	private double speed;
 	
-	private CyclistType(int speed) {
+	private CyclistType(double speed) {
 		this.speed = speed;
 	}
 	
-	public int getCyclistSpeed() {
+	public double getCyclistSpeed() {
 		return this.speed;
 	}
 }
